area y encargado MongoDB

+++ routes/areaRoutes.js
@@ -3,27 +3,27 @@ const router = express.Router();
 const areasService = require('../services/areaService');
 const service = new areasService();
 
-router.get('/', (req, res) => {
+router.get('/', async (req, res) => {
   const { id } = req.query;
-  const areas = id ? service.getById(id) : service.getAll();
+  const areas = id ? await service.getById(id) : await service.getAll();
   if (areas) {
     res.status(200).json(areas);
   } else {
-    res.status(404).json({ message: 'Error ' });
+    res.status(404).json({ message: 'No se encontraron áreas ' });
   }
 });
 
-router.post('/', (req, res) => {
+router.post('/', async (req, res) => {
   const body = req.body;
   try {
-    service.create(body);
-    res.status(200).json({ message: 'Area creada', data: body });
+    const newArea = await service.create(body);
+    res.status(200).json({ message: 'Area creada', data: newArea });
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
 });
 
-router.patch('/', (req, res) => {
+router.patch('/', async (req, res) => {
   const { id } = req.query;
   const datosActualizados = req.body;
 
@@ -32,7 +32,7 @@ router.patch('/', (req, res) => {
   }
 
   try {
-    const areaActualizada = service.update(id, datosActualizados);
+    const areaActualizada = await service.update(id, datosActualizados);
     res
       .status(200)
       .json({ message: 'Área actualizada', data: areaActualizada });
@@ -49,10 +49,15 @@ router.patch('/', (req, res) => {
   }
 });
 
-router.delete('/', (req, res) => {
+router.delete('/', async (req, res) => {
   const { id } = req.query;
+
+  if (!id) {
+    return res.status(400).json({ message: 'El parámetro id es requerido.' });
+  }
+
   try {
-    service.delete(id);
+    await service.delete(id);
     res.status(200).json({ message: 'El área fue eliminada correctamente ' });
   } catch (error) {
     if (error.message.includes('no existe')) {
